--- crates/polars-ops/src/chunked_array/datetime/mod.rs
@@ -1,4 +1,0 @@
-#[cfg(feature = "timezones")]
-mod replace_time_zone;
-#[cfg(feature = "timezones")]
-pub use replace_time_zone::*;
